requirements.txt
flask==1.1.2
werkzeug==1.0.1
MarkupSafe==1.1.1
Biopython